phan quyen category

+++ src/main/java/com/example/web_shop_ptit/admin/controller/CategoryAdminController.java
@@ -1,7 +1,10 @@
 package com.example.web_shop_ptit.admin.controller;
 
+import com.example.web_shop_ptit.admin.entity.Admin;
 import com.example.web_shop_ptit.admin.entity.CategoryManagement;
 import com.example.web_shop_ptit.admin.service.CategoryManagementService;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
 import org.hibernate.engine.jdbc.mutation.spi.BindingGroup;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,6 +17,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.util.List;
+import java.util.Objects;
 
 @Controller
 @RequestMapping("/admin/")
@@ -23,11 +27,13 @@ public class CategoryAdminController {
     private CategoryManagementService categoryManagementService;
 
     @GetMapping("/category")
-    public String categoryPage(Model model) {
+    public String categoryPage(Model model, HttpServletRequest request) {
         List<CategoryManagement> categories = categoryManagementService.listAll();
 
         model.addAttribute("categories", categories);
-
+        HttpSession session = request.getSession();
+        Admin admin = (Admin) session.getAttribute("adminInfo");
+        model.addAttribute("role", admin.getRole());
         // Trả về tên của view hoặc đường dẫn
         return "web_admin/category";
     }
@@ -78,7 +84,13 @@ public class CategoryAdminController {
         }
     }
     @RequestMapping("addCategory")
-    public String addCategoryPage(Model model) {
+    public String addCategoryPage(Model model, HttpServletRequest request, RedirectAttributes redirectAttributes) {
+        HttpSession session = request.getSession();
+        Admin admin = (Admin) session.getAttribute("adminInfo");
+        if(!Objects.equals(admin.getRole(), "admin") && !Objects.equals(admin.getRole(), "manager")){
+            redirectAttributes.addFlashAttribute("err", "Bạn không có quyền thực hiện thao tác này!");
+            return "redirect:/admin/category";
+        }
         List<CategoryManagement> categories = categoryManagementService.listAll();
         model.addAttribute("categories", categories);
         model.addAttribute("operation", "add");
@@ -90,8 +102,14 @@ public class CategoryAdminController {
     public String updateCategoryPage(@RequestParam int maDanhMuc,
                                      @RequestParam String tenDanhMuc,
                                      @RequestParam int danhMucCha,
-                                     Model model) {
+                                     Model model, HttpServletRequest request, RedirectAttributes redirectAttributes) {
         try {
+            HttpSession session = request.getSession();
+            Admin admin = (Admin) session.getAttribute("adminInfo");
+            if(!Objects.equals(admin.getRole(), "admin") && !Objects.equals(admin.getRole(), "manager")){
+                redirectAttributes.addFlashAttribute("err", "Bạn không có quyền thực hiện thao tác này!");
+                return "redirect:/admin/category";
+            }
             CategoryManagement category = new CategoryManagement();
             category.setCategoryId(maDanhMuc);
             category.setName(tenDanhMuc);
@@ -111,7 +129,13 @@ public class CategoryAdminController {
     }
 
     @RequestMapping("deleteCategory")
-    public String deleteCategoryPage(@RequestParam int categoryId,RedirectAttributes redirectAttributes) {
+    public String deleteCategoryPage(@RequestParam int categoryId,RedirectAttributes redirectAttributes, HttpServletRequest request) {
+        HttpSession session = request.getSession();
+        Admin admin = (Admin) session.getAttribute("adminInfo");
+        if(!Objects.equals(admin.getRole(), "admin") && !Objects.equals(admin.getRole(), "manager")){
+            redirectAttributes.addFlashAttribute("err", "Bạn không có quyền thực hiện thao tác này!");
+            return "redirect:/admin/category";
+        }
         categoryManagementService.deleteCategory(categoryId);
         redirectAttributes.addFlashAttribute("success", "Xóa danh mục thành công");
         return "redirect:/admin/category";
